+++ sdk/metric/view_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 
 	"github.com/go-logr/logr"
+	"github.com/go-logr/logr/funcr"
 	"github.com/go-logr/logr/testr"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -405,6 +406,18 @@ func TestNewViewReplace(t *testing.T) {
 			},
 		},
 		{
+			name: "AttributeKeys",
+			mask: Stream{AllowAttributeKeys: []attribute.Key{"test"}},
+			want: func(i Instrument) Stream {
+				return Stream{
+					Name:               i.Name,
+					Description:        i.Description,
+					Unit:               i.Unit,
+					AllowAttributeKeys: []attribute.Key{"test"},
+				}
+			},
+		},
+		{
 			name: "Complete",
 			mask: Stream{
 				Name:        alt,
@@ -430,23 +443,6 @@ func TestNewViewReplace(t *testing.T) {
 			assert.Equal(t, test.want(completeIP), got)
 		})
 	}
-
-	// Go does not allow for the comparison of function values, even their
-	// addresses. Therefore, the AttributeFilter field needs an alternative
-	// testing strategy.
-	t.Run("AttributeFilter", func(t *testing.T) {
-		allowed := attribute.String("key", "val")
-		filter := func(kv attribute.KeyValue) bool {
-			return kv == allowed
-		}
-		mask := Stream{AttributeFilter: filter}
-		got, match := NewView(completeIP, mask)(completeIP)
-		require.True(t, match, "view did not match exact criteria")
-		require.NotNil(t, got.AttributeFilter, "AttributeFilter not set")
-		assert.True(t, got.AttributeFilter(allowed), "wrong AttributeFilter")
-		other := attribute.String("key", "other val")
-		assert.False(t, got.AttributeFilter(other), "wrong AttributeFilter")
-	})
 }
 
 type badAgg struct {
@@ -469,6 +465,30 @@ func TestNewViewAggregationErrorLogged(t *testing.T) {
 	require.True(t, match, "view did not match exact criteria")
 	assert.Nil(t, got.Aggregation, "erroring aggregation used")
 	assert.Equal(t, 1, l.ErrorN())
+}
+
+func TestNewViewEmptyViewErrorLogged(t *testing.T) {
+	var got string
+	otel.SetLogger(funcr.New(func(_, args string) {
+		got = args
+	}, funcr.Options{Verbosity: 6}))
+
+	_ = NewView(Instrument{}, Stream{})
+	assert.Contains(t, got, errEmptyView.Error())
+}
+
+func TestNewViewMultiInstMatchErrorLogged(t *testing.T) {
+	var got string
+	otel.SetLogger(funcr.New(func(_, args string) {
+		got = args
+	}, funcr.Options{Verbosity: 6}))
+
+	_ = NewView(Instrument{
+		Name: "*", // Wildcard match name (multiple instruments).
+	}, Stream{
+		Name: "non-empty",
+	})
+	assert.Contains(t, got, errMultiInst.Error())
 }
 
 func ExampleNewView() {